chiply-isr/server/pyproject.toml
[project]
name = "org-db-server"
version = "0.1.0"
description = "FastAPI server for org-db v3 with semantic search, image search, and full-text search"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "John Kitchin", email = "jkitchin@andrew.cmu.edu" }
]
keywords = [
    "org-mode",
    "emacs",
    "semantic-search",
    "vector-search",
    "knowledge-management",
    "fastapi",
    "embeddings",
    "clip",
    "image-search"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Framework :: FastAPI",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=1.8.2",
    "pymupdf4llm>=0.0.5",
    "python-docx>=1.0.0",
    "python-pptx>=0.6.0",
    "markitdown[all]>=0.0.1",
    "psutil>=5.9.0",
    "libsql>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]
